services/air-quality-service/requirements-prod.txt
aiohttp==3.9.1
python-dotenv==1.0.0
influxdb3-python==0.3.0

